Tiger

+++ src/main/java/team/snof/simplesearch/search/model/dao/index/TempData.java
@@ -1,6 +1,7 @@
 package team.snof.simplesearch.search.model.dao.index;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -9,5 +10,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class TempData {
     private long docId;
-    private long wordFreq;
+    private Integer wordFreq;
 }
